Don't add D2/He3 to SpaceDust scanner/analyzer if FFT is present
Nertea opted to add D2 to the SpaceDust telescope if FFT is present, but not the SpaceDust scanner or SpaceDust analyzer. Instead he added a specialized scanner to FFT with D2/He3 capabilities, which I take to mean that he meant those to be exclusive to that part, much like how only the antimatter scanner can detect antimatter.